Lora

+++ controllers/beachesController.js
@@ -3,50 +3,47 @@ const db = require("../models");
 const router = express.Router();
 const methodOverride = require("method-override");
 
-// ----- Routes ----- //
 
-// NEW
+// NEW BEACH PAGE
 router.get("/new", (req, res) => {
+  if(!req.session.currentUser) return res.redirect('/'); 
   db.COUNTY.find({}, (err, counties) => {
     if (err) console.log(err);
     res.render("new", { counties });
   });
 });
 
-// ADD
+// ADD BEACH
 router.post("/", (req, res) => {
+  if(!req.session.currentUser) return res.redirect('/');
   db.BEACH.create(req.body, (err, newBeach) => {
     if (err) console.log(err);
     db.COUNTY.findById(req.body.countyId, (err, foundCounty) => {
       if (err) console.log(err);
-      foundCounty.beaches.push(newBeach);
-      foundCounty.save((err, savedCounty) => {
-        if (err) console.log(err);
-        console.log("savedCounty:", savedCounty);
-        res.redirect("/beaches");
+      newBeach.county.push(foundCounty);
+      newBeach.save((err, newBeach)=> {
+        if(err) console.log(err);
+        foundCounty.beaches.push(newBeach);
+        foundCounty.save((err, savedCounty) => {
+          if (err) console.log(err);
+          console.log("savedCounty:", savedCounty);
+          res.redirect("/beaches");
+        });
       });
     });
   });
 });
 
-// SHOW
+// SHOW BEACH
 router.get("/:id", (req, res) => {
-  db.BEACH.findById(req.params.id, (err, showBeach) => {
-    if (err) return console.log(err);
-    db.COMMENT.find({}, (err, allComments) => {
+  db.BEACH.findById(req.params.id)
+  .populate('comments')
+  .populate('county') // did 2 populates, IT
+  .exec((err, foundBeach) => {
       if (err) console.log(err);
-      // console.log(allComments); // this currently shows all comments for all beaches
-    })
-      .populate({
-        path: "beach",
-        match: { _id: req.params.id },
-      })
-      .exec((err, allComments) => {
-        if (err) console.log(err);
-        res.render("show", {
-          beach: showBeach,
-          comments: allComments,
-        });
+      console.log(foundBeach);
+      res.render("show", {
+          beach: foundBeach,
       });
   });
 });
@@ -61,8 +58,9 @@ router.get("/", (req, res) => {
   });
 });
 
-// EDIT
+// EDIT ///////////////////// NOT WORKING ///////////////////// 
 router.get("/:id/edit", (req, res) => {
+  if(!req.session.currentUser) return res.redirect('/'); 
   db.COUNTY.find({}, (err, allCounties) => {
     db.COUNTY.findOne({ beaches: req.params.id })
       .populate({
@@ -80,8 +78,9 @@ router.get("/:id/edit", (req, res) => {
   });
 });
 
-// UPDATE
+// UPDATE ///////////////////// NOT WORKING /////////////////////
 router.put("/:id", (req, res) => {
+  if(!req.session.currentUser) return res.redirect('/'); // this should redirect to homepage if user is not logged in and tries to create a new beach
   db.BEACH.findByIdAndUpdate(
     req.params.id, // find by id
     req.body, // update by id
@@ -109,6 +108,7 @@ router.put("/:id", (req, res) => {
 
 // DELETE
 router.delete("/:id", (req, res) => {
+  if(!req.session.currentUser) return res.redirect('/'); 
   db.BEACH.findByIdAndDelete(req.params.id, (err, deleteBeach) => {
     if (err) return console.log(err);
     console.log("Deleted Beach:", deleteBeach);
@@ -116,12 +116,12 @@ router.delete("/:id", (req, res) => {
   });
 });
 
-// ----- Export Controller ----- //
 
-// ----------------- Comments
+// ----------------- Comments ----------------- //
 
-// NEW COMMENT FORM ON BEACHES SHOW PAGE
+// NEW COMMENT FROM BEACH SHOW PAGE
 router.get("/:beachid/comments/new", (req, res) => {
+  if(!req.session.currentUser) return res.redirect('/'); 
   db.BEACH.findById(req.params.id, (err, foundBeach) => {
     console.log(foundBeach);
     if (err) console.log(err);
@@ -131,6 +131,7 @@ router.get("/:beachid/comments/new", (req, res) => {
 
 // CREATE NEW COMMENT AND PUSH TO BEACH.COMMENTS ARRAY
 router.post("/:beachid/comments/new", (req, res) => {
+  if(!req.session.currentUser) return res.redirect('/'); // this should redirect to homepage if user is not logged in and tries to create a new beach
   db.COMMENT.create(req.body, (err, newComment) => {
     if (err) return console.log(err);
     console.log(req.params.beachid);
@@ -146,6 +147,7 @@ router.post("/:beachid/comments/new", (req, res) => {
   });
 });
 
-// ----- Export Controller ----- //
+
+
 
 module.exports = router;
